requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
attrs==21.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
boto3==1.17.91
botocore==1.20.91; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
click==8.0.1; python_version >= '3.6'
flask-sqlalchemy==2.5.1
flask==2.0.1
greenlet==1.1.0; python_version >= '3'
iniconfig==1.1.1
itsdangerous==2.0.1; python_version >= '3.6'
jinja2==3.0.1; python_version >= '3.6'
jmespath==0.10.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
joblib==1.0.1; python_version >= '3.6'
markupsafe==2.0.1; python_version >= '3.6'
numpy==1.21.0rc2
packaging==20.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pandas==1.2.4
pluggy==1.0.0.dev0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
py==1.10.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pymysql==1.0.2
pyparsing==3.0.0b2; python_version >= '3.5'
pytest==6.2.4
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2021.1
pyyaml==5.4.1
s3transfer==0.4.2
scikit-learn==0.24.2; python_version >= '3.6'
scipy==1.7.0rc1; python_version < '3.10' and python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sklearn==0.0
sqlalchemy==1.4.17
threadpoolctl==2.1.0; python_version >= '3.5'
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
urllib3==1.26.5; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
werkzeug==2.0.1; python_version >= '3.6'
